services/course_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
flask-swagger-ui==4.11.1